restructured setting and update pages
