+++ src/Routes.jsx
@@ -6,11 +6,11 @@ import Login from "./pages/Login";
 import MyAccount from "./pages/MyAccount";
 import NoPage from "./pages/NoPage";
 import ProtectRoutes from "./ProtectRoutes";
-import Settings from "./pages/Settings";
+import Settings from "./pages/SettingsPage/Settings";
 import SignUp from "./pages/SignUp";
-import UpdateEmail from "./pages/UpdateEmail";
-import UpdatePassword from "./pages/UpdatePassword";
-import UpdateUsername from "./pages/UpdateUsername";
+import UpdateEmail from "./pages/UpdatePages/UpdateEmail";
+import UpdatePassword from "./pages/UpdatePages/UpdatePassword";
+import UpdateUsername from "./pages/UpdatePages/UpdateUsername";
 import UploadImage from "./pages/UploadImage";
 import VerifyNewUsers from "./pages/VerifyNewUsersPage/VerifyNewUsers";
 import NewTodo from "./pages/todoToMake";
