+++ src/pages/UpdatePages/_update.scss
@@ -1,0 +1,23 @@
+.update-user-info {
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: initial;
+  gap: 2rem;
+  align-items: center;
+  label {
+    @include flex(row nowrap);
+    width: 100%;
+  }
+  .update-back {
+    width: 100%;
+    padding: 1rem;
+  }
+  .update-password {
+    height: 90%;
+  }
+  button {
+    border: none;
+    border-radius: 5px;
+    background: transparent;
+  }
+}
