+++ src/styles/index.scss
@@ -17,10 +17,13 @@ body {
 
 //animated toasifypop
 @import "./toasifypop";
+@import "./back";
 
 //Pages styles
 @import "./pages/HomePage/home";
 @import "./pages/ForgottenPasswordPage/verify";
+@import "./pages/UpdatePages/update";
+@import "./pages/SettingsPage/settings";
 
 //dark/light mode switch styles
 .main-container-relative {
@@ -552,121 +555,121 @@ body {
 //     }
 //   }
 // }
-.update-user-info {
-  display: flex;
-  flex-flow: column nowrap;
-  justify-content: initial;
-  gap: 2rem;
-  align-items: center;
-  label {
-    @include flex(row nowrap);
-    width: 100%;
-  }
-  .update-back {
-    width: 100%;
-    padding: 1rem;
-  }
-  .update-password {
-    height: 90%;
-  }
-  button {
-    border: none;
-    border-radius: 5px;
-    background: transparent;
-  }
-}
+// .update-user-info {
+//   display: flex;
+//   flex-flow: column nowrap;
+//   justify-content: initial;
+//   gap: 2rem;
+//   align-items: center;
+//   label {
+//     @include flex(row nowrap);
+//     width: 100%;
+//   }
+//   .update-back {
+//     width: 100%;
+//     padding: 1rem;
+//   }
+//   .update-password {
+//     height: 90%;
+//   }
+//   button {
+//     border: none;
+//     border-radius: 5px;
+//     background: transparent;
+//   }
+// }
 
-.back {
-  button {
-    @include flex(row nowrap);
-    width: 50px;
-    height: 50px;
-    border: none;
-    background: $button;
-    border-radius: 50%;
-    margin-bottom: 1rem;
-    a {
-      @include flex(row nowrap);
-      border: inherit;
-      width: 100%;
-      height: 100%;
-      border-radius: 50%;
-      background: transparent;
-      font-size: 1.7rem;
-      color: var(--black, $black);
-      text-decoration: none;
-      .fa-solid {
-        background: transparent;
-      }
-    }
-  }
-  button {
-    transition: transform 1s;
-    cursor: pointer;
-  }
-  button:hover {
-    transform: scale(1.1);
-  }
-  button:active {
-    transform: scale(0.5);
-  }
-}
+// .back {
+//   button {
+//     @include flex(row nowrap);
+//     width: 50px;
+//     height: 50px;
+//     border: none;
+//     background: $button;
+//     border-radius: 50%;
+//     margin-bottom: 1rem;
+//     a {
+//       @include flex(row nowrap);
+//       border: inherit;
+//       width: 100%;
+//       height: 100%;
+//       border-radius: 50%;
+//       background: transparent;
+//       font-size: 1.7rem;
+//       color: var(--black, $black);
+//       text-decoration: none;
+//       .fa-solid {
+//         background: transparent;
+//       }
+//     }
+//   }
+//   button {
+//     transition: transform 1s;
+//     cursor: pointer;
+//   }
+//   button:hover {
+//     transform: scale(1.1);
+//   }
+//   button:active {
+//     transform: scale(0.5);
+//   }
+// }
 
-.settings,
-.my-account {
-  display: flex;
-  flex-flow: column nowrap;
-  gap: 1rem;
-  padding: 1rem 2rem;
-  background: var(--white, $white);
-  width: 100vw;
-  height: 100vh;
-  font-size: 1.2rem;
+// .settings,
+// .my-account {
+//   display: flex;
+//   flex-flow: column nowrap;
+//   gap: 1rem;
+//   padding: 1rem 2rem;
+//   background: var(--white, $white);
+//   width: 100vw;
+//   height: 100vh;
+//   font-size: 1.2rem;
 
-  .light-dark-mode,
-  .updates {
-    @include border();
-    color: var(--black, $black);
-    border-radius: 10px;
-    height: 15%;
-    padding: 0.5rem 1rem;
-    display: flex;
-    flex-flow: row nowrap;
-    justify-content: space-between;
-    align-items: center;
-    .lists {
-      width: 100%;
-      height: 100%;
-      text-align: left;
-      background: transparent;
-      border: none;
-      color: inherit;
-      font-size: 1.2rem;
-      cursor: pointer;
-      a {
-        display: flex;
-        align-items: center;
-        text-decoration: none;
-        color: inherit;
-        width: 100%;
-        height: 100%;
-        padding: 1rem;
-      }
-    }
-  }
-  .updates {
-    transition: transform 1s;
-  }
-  .updates:hover {
-    transform: scale(1.03);
-  }
-  .updates:active {
-    transform: scale(0.8);
-  }
-  .delete {
-    color: $red;
-  }
-}
+//   .light-dark-mode,
+//   .updates {
+//     @include border();
+//     color: var(--black, $black);
+//     border-radius: 10px;
+//     height: 15%;
+//     padding: 0.5rem 1rem;
+//     display: flex;
+//     flex-flow: row nowrap;
+//     justify-content: space-between;
+//     align-items: center;
+//     .lists {
+//       width: 100%;
+//       height: 100%;
+//       text-align: left;
+//       background: transparent;
+//       border: none;
+//       color: inherit;
+//       font-size: 1.2rem;
+//       cursor: pointer;
+//       a {
+//         display: flex;
+//         align-items: center;
+//         text-decoration: none;
+//         color: inherit;
+//         width: 100%;
+//         height: 100%;
+//         padding: 1rem;
+//       }
+//     }
+//   }
+//   .updates {
+//     transition: transform 1s;
+//   }
+//   .updates:hover {
+//     transform: scale(1.03);
+//   }
+//   .updates:active {
+//     transform: scale(0.8);
+//   }
+//   .delete {
+//     color: $red;
+//   }
+// }
 .loading {
   @include flex(row nowrap);
   border: 10px solid #f3f3f3;
